fix: types handling in electron
