+++ apps/electron/package.json
@@ -36,7 +36,6 @@
     "@types/better-sqlite3": "^7.6.4",
     "@types/fs-extra": "^11.0.1",
     "cross-env": "7.0.3",
-    "dts-for-context-bridge": "^0.7.1",
     "electron": "24.1.2",
     "electron-log": "^5.0.0-beta.23",
     "electron-squirrel-startup": "1.0.0",
